requirements.txt
click
faiss-gpu
numpy
opencv-python
opencv-contrib-python
pillow
scipy
tqdm
